categoria service layer validations

+++ src/main/java/com/example/demo/domain/services/impl/CategoriaDomainServiceImpl.java
@@ -9,21 +9,31 @@ import org.springframework.stereotype.Service;
 
 import com.example.demo.application.dtos.CategoriaRequestDto;
 import com.example.demo.application.dtos.CategoriaResponseDto;
+import com.example.demo.domain.exceptions.CategoriaComNomeDuplicadoException;
 import com.example.demo.domain.models.entities.Categoria;
 import com.example.demo.domain.services.interfaces.CategoriaDomainService;
 import com.example.demo.infrastructure.repositories.CategoriaRepository;
+import com.example.demo.infrastructure.repositories.ProdutoRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class CategoriaDomainServiceImpl implements CategoriaDomainService {
 
 	@Autowired
 	private CategoriaRepository categoriaRepository;
+	
+	@Autowired
+	private ProdutoRepository produtoRepository;
 
 	@Autowired
 	private ModelMapper modelMapper;
 
 	@Override
 	public String registrarCategoria(CategoriaRequestDto request) {
+
+		if (categoriaRepository.existsByNome(request.getNome()))
+			throw new CategoriaComNomeDuplicadoException(request.getNome());
 
 		var categoria = new Categoria();
 		categoria.setId(UUID.randomUUID());
@@ -37,6 +47,12 @@ public class CategoriaDomainServiceImpl implements CategoriaDomainService {
 	@Override
 	public CategoriaResponseDto editarCategoria(UUID id, CategoriaRequestDto request) {
 
+		if (!categoriaRepository.existsById(id))
+			throw new EntityNotFoundException("Categoria com ID " + id + " não encontrada.");
+
+		if (categoriaRepository.existsByNome(request.getNome()))
+			throw new CategoriaComNomeDuplicadoException(request.getNome());
+
 		var categoria = categoriaRepository.findById(id).get();
 		categoria.setNome(request.getNome());
 
@@ -47,6 +63,12 @@ public class CategoriaDomainServiceImpl implements CategoriaDomainService {
 
 	@Override
 	public String excluirCategoria(UUID id) {
+		
+		if (!categoriaRepository.existsById(id))
+			throw new EntityNotFoundException("Categoria com ID " + id + " não encontrada.");
+		
+		if (produtoRepository.existsByCategoriaId(id))
+	        throw new RuntimeException("Não é possível excluir a categoria ID " + id + " pois existem produtos associados a ela.");
 
 		categoriaRepository.deleteById(id);
 
@@ -55,7 +77,10 @@ public class CategoriaDomainServiceImpl implements CategoriaDomainService {
 
 	@Override
 	public CategoriaResponseDto consultarCategoriaPorId(UUID id) {
-
+		
+		if (!categoriaRepository.existsById(id))
+			throw new EntityNotFoundException("Categoria com ID " + id + " não encontrada.");
+	    
 		return modelMapper.map(categoriaRepository.findById(id).get(), CategoriaResponseDto.class);
 	}
